+++ src/http/data_loader.rs
@@ -11,7 +11,7 @@ use crate::config::Batch;
 use crate::data_loader::{DataLoader, Loader};
 use crate::http::{DataLoaderRequest, Response};
 use crate::json::JsonLike;
-use crate::target_runtime::TargetRuntime;
+use crate::runtime::TargetRuntime;
 
 fn get_body_value_single(body_value: &HashMap<String, Vec<&ConstValue>>, id: &str) -> ConstValue {
     body_value
