--- tests/graphql_spec.graphql
@@ -1,1 +1,0 @@
-directive @error(message: String!, trace: [String!]!, description: String) repeatable on OBJECT
